Separated PointOfinterest model to other files
